src/main/java/com/liushuo/wsell/Service/Impl/CategoryServiceImpl.java
package com.liushuo.wsell.Service.Impl;

import com.liushuo.wsell.Repository.ProductCategoryRepository;
import com.liushuo.wsell.Service.CategoryService;
import com.liushuo.wsell.dataobject.ProductCategory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Create by liushuo on 2017/11/27.
 */
@Service
public class CategoryServiceImpl implements CategoryService {
    @Autowired
    private ProductCategoryRepository productCategoryRepository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return productCategoryRepository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryRepository.findAll();
    }

    @Override
    public List<ProductCategory> findCategorysByType(List<Integer> list) {
        return productCategoryRepository.findAllByCategoryTypeIn(list);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryRepository.save(productCategory);
    }
}
